create components class
